Import FQCNs
`import_symbols` by default is set to `false`. If we want to make sure that we're always importing FQCNs, we need to enable it.

See https://cs.symfony.com/doc/rules/import/fully_qualified_strict_types.html#example-4

This is an important rule when used in combination with tools like Rector that are always using FQCNs for the code rewriting process.